Fix bug in get_cropping_bounds for images
Could sometimes give incorrect crop on bottom row or right-most column.
This was just an off-by-1 error in the indexing.  This could actually
end up mattering a lot because at the edges of the image one can end up
with blank columns when combined with cv2.connetedComponents, as the
definition of "connected" includes pixels diagonally attached.